ZNTA-1235 - show confirmation before deleting

+++ frontend/src/main/web/lib/actions/index.js
@@ -180,19 +180,25 @@ export function getWorkDetail(id) {
   }
 }
 
-// =========== delete a single work
+export const TOGGLE_DELETE_CONFIRMATION = 'TOGGLE_DELETE_CONFIRMATION'
+export const toggleDeleteConfirmation = createAction(TOGGLE_DELETE_CONFIRMATION)
+
+// =========== delete a single work config
 export const DELETE_WORK_REQUEST = 'DELETE_WORK_REQUEST'
 export const DELETE_WORK_SUCCESS = 'DELETE_WORK_SUCCESS'
 export const DELETE_WORK_FAILURE = 'DELETE_WORK_FAILURE'
-export function deleteWork(id) {
+export function deleteWork() {
   return (dispatch, getState) => {
-    dispatch({
-      [CALL_API]: {
-        endpoint: `${getState().configs.apiUrl}/api/work/${id}`,
-        method: 'DELETE',
-        credentials: 'include',
-        types: [DELETE_WORK_REQUEST, DELETE_WORK_SUCCESS, DELETE_WORK_FAILURE]
-      }
-    })
+    const selectedConfig = getState().workConfig.workDetail
+    if (selectedConfig) {
+      dispatch({
+        [CALL_API]: {
+          endpoint: `${getState().configs.apiUrl}/api/work/${selectedConfig.id}`,
+          method: 'DELETE',
+          credentials: 'include',
+          types: [DELETE_WORK_REQUEST, DELETE_WORK_SUCCESS, DELETE_WORK_FAILURE]
+        }
+      })
+    }
   }
 }
